feat: 통계, 점주계정CRUD, 관리자 대시보드

+++ src/main/java/org/example/luckyburger/domain/shop/dto/response/ShopDashboardResponse.java
@@ -4,12 +4,12 @@ import lombok.Builder;
 
 @Builder
 public record ShopDashboardResponse(
-        Integer todayOrderCount,
+        Long todayOrderCount,
         Long todayTotalSales,
         Double averageRating
 ) {
     public static ShopDashboardResponse of(
-            Integer todayOrderCount,
+            Long todayOrderCount,
             Long todayTotalSales,
             Double averageRating) {
         return ShopDashboardResponse.builder()
